FISH-9771 move random utils to common-util
