+++ nucleus/common/common-util/src/main/java/org/glassfish/common/util/RandomUtils.java
@@ -1,23 +1,23 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2010 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2024 Payara Foundation and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
  * and Distribution License("CDDL") (collectively, the "License").  You
  * may not use this file except in compliance with the License.  You can
  * obtain a copy of the License at
- * https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
- * or packager/legal/LICENSE.txt.  See the License for the specific
+ * https://github.com/payara/Payara/blob/main/LICENSE.txt
+ * See the License for the specific
  * language governing permissions and limitations under the License.
  *
  * When distributing the software, include this License Header Notice in each
- * file and include the License file at packager/legal/LICENSE.txt.
+ * file and include the License file at glassfish/legal/LICENSE.txt.
  *
  * GPL Classpath Exception:
- * Oracle designates this particular file as subject to the "Classpath"
- * exception as provided by Oracle in the GPL Version 2 section of the License
+ * The Payara Foundation designates this particular file as subject to the "Classpath"
+ * exception as provided by the Payara Foundation in the GPL Version 2 section of the License
  * file that accompanied this code.
  *
  * Modifications:
@@ -37,48 +37,23 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-// Portions Copyright [2019] Payara Foundation and/or affiliates
-
-package com.sun.enterprise.universal.security;
+package org.glassfish.common.util;
 
 import java.security.SecureRandom;
 
 /**
- * Created October 28, 2010
- * @author Byron Nevins
+ * Centralize usage of random number generator.
+ *
+ * @author: Petr Aubrecht
  */
-public final class SecurityUtils {
+public class RandomUtils {
     private static SecureRandom random = new SecureRandom();
 
-    public static String getSecureRandomHexString(int numBytes) {
-        byte[] bb = new byte[numBytes];
-        random.nextBytes(bb);
-        return toHexString(bb);
+    public static int nextInt() {
+        return random.nextInt();
     }
 
-    /**
-     * No instances allowed.
-     */
-    private SecurityUtils() {
-    }
-
-    private static String toHexString(byte b) {
-        String s = Integer.toHexString((int) b + 128);
-
-        if (s.length() == 1) {
-            s = "0" + s;
-        }
-
-        return s;
-    }
-
-    private static String toHexString(byte[] bb) {
-        StringBuilder sb = new StringBuilder();
-
-        for (byte b : bb) {
-            sb.append(toHexString(b));
-        }
-
-        return sb.toString();
+    public static int nextInt(int bound) {
+        return random.nextInt(bound);
     }
 }
